requirements.txt
flask
pandas
scikit-learn
duckduckgo-search
newspaper3k
easyocr
torch
opencv-python
numpy
